cleanup old stupid tests
