--- test/stack_test.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "../src/runtime/memory/gc.h"
-
-#include <stdio.h>
-#include <stdlib.h>
-
-int test_stack_walk();
